% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rif.R
\name{est_rif}
\alias{est_rif}
\title{Estimate Recentered Influence Function}
\usage{
est_rif(
  functional,
  dep_var,
  weights = NULL,
  quantiles = NULL,
  custom_rif_function = NULL,
  ...
)
}
\arguments{
\item{functional}{string containing the functional for which to compute the RIF. Can be one of
"mean", "variance", "quantiles", "gini", or "custom". If "custom"
is selected a \code{custom_rif_function} needs to be provided.}

\item{dep_var}{dependent variable of distributional function. Discrete or continuous numeric vector.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{dep_var}.
The default (\code{NULL)} is equivalent to \code{weights = rep(1/nx, nx)},
where nx is the length of (the finite entries of) \code{dep_var}.}

\item{quantiles}{a vector of length 1 or more with quantile positions to calculate the RIF.
Each quantile is indicated with value between 0 and 1.}

\item{custom_rif_function}{the RIF function to compute the RIF of the custom functional.}

\item{...}{the parameters passed to the \code{custom_rif_function}. The must have a different name
than the the ones of \code{est_rif}. For instance, if you want to pass weights to the
\code{custom_rif_function}, name them \code{custom_weights}.}
}
\value{
a data frame with the RIF value for each observation and in the case of several quantiles
        a column for each quantile.
}
\description{
This function estimates the recentered influence function (RIF) of a chosen functional
(e.g. quantiles, variance or gini).
}
