% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rif.R
\name{est_rif_mean}
\alias{est_rif_mean}
\title{Estimate RIF of the Mean}
\usage{
est_rif_mean(dep_var)
}
\arguments{
\item{dep_var}{dependent variable of distributional function. Discrete or continuous numeric vector.}
}
\value{
A data frame with the number of columns equaling the length of vector \code{probs}. Each column contains the RIF values of the probs.
}
\description{
Function to estimate the recentered influence function (RIF) of the mean
of a weighted distribution of a dependent variable.
}
\examples{

dep_var <- c(1, 3, 9, 16, 3, 7, 4, 9)
est_rif_mean(dep_var)

}
