% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifreg.R
\name{est_rif_of_mean}
\alias{est_rif_of_mean}
\title{Estimate RIF of Mean}
\usage{
est_rif_of_mean(dep_var, weights)
}
\arguments{
\item{dep_var}{dependent variable of distributional function. Discrete or continous numeric vector.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{dep_var}.
The default NULL is equivalent to weights = rep(1/nx, nx) where nx is the length of (the finite entries of) dep_var[].}
}
\value{
A data frame with a single column containing the RIF values of the mean.
}
\description{
Function to estimate the recentered influence function (RIF) of the mean
of a weighted distribution of a dependent variable.
}
