% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifreg.R
\name{est_rif_of_quantiles}
\alias{est_rif_of_quantiles}
\alias{est_rif_of_quantile}
\title{Estimate RIF of Quantiles}
\usage{
est_rif_of_quantiles(quantiles, dep_var, weights = NULL, ...)

est_rif_of_quantile(quantile, dep_var, weights, density)
}
\arguments{
\item{quantiles}{quantile position to calculate the RIF. Vector of length 1 or more containing positive doubles with value < 1.}

\item{dep_var}{dependent variable of distributional function. Discrete or continous numeric vector.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{dep_var}.
The default NULL is equivalent to weights = rep(1/nx, nx) where nx is the length of (the finite entries of) dep_var[].}

\item{...}{further arguments passed on to \code{stats::density()} function.}

\item{quantile}{specific quantile at which to estimate the RIF values. Positive double with value < 1.}

\item{density}{kernel density approximation of \code{dep_var} calculated using stats::density().}
}
\value{
A data frame with the number of columns equaling the length of quantities[]. Each column contains the RIF values at the quantiles.
}
\description{
Function to estimate the recentered influence function (RIF) of one or several specified quantiles
of a weighted distribution of a dependent variable.
}
\section{Functions}{
\itemize{
\item \code{est_rif_of_quantile}: helper function to estimate the RIF values of a specific quantile.
}}

\examples{
dep_var <- c(1, 3, 9, 16, 3, 7, 4, 9)
quantiles <- seq(1:9)/10
weights <- c(2, 1, 3, 4, 4, 1, 6, 3)
est_rif_of_quantiles(quantiles, dep_var, weights)

}
