% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rif.R
\name{est_rif_quantiles}
\alias{est_rif_quantiles}
\title{Estimate RIF of Quantiles}
\usage{
est_rif_quantiles(dep_var, quantiles, weights = NULL, ...)
}
\arguments{
\item{quantiles}{quantile position to calculate the RIF. Positive double with value < 1.}

\item{...}{further arguments passed on to \code{stats::density()} function.}
}
\value{
A data frame with the number of columns equaling the length of vector \code{quantiles}. Each column contains the RIF values of the quantiles.
}
\description{
Function to estimate the recentered influence function (RIF) of one or several specified quantiles
of a weighted distribution of a dependent variable.
}
\examples{
dep_var <- c(1, 3, 9, 16, 3, 7, 4, 9)
quantiles <- seq(1:9)/10
weights <- c(2, 1, 3, 4, 4, 1, 6, 3)
est_rif_at_quantiles(dep_var, quantiles, weights = NULL)

}
