% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifreg.R
\name{est_rifreg}
\alias{est_rifreg}
\title{Estimate RIF Regression}
\usage{
est_rifreg(
  formula,
  data,
  functional,
  custom_rif_function = NULL,
  quantiles = NULL,
  weights = NULL,
  bootstrap = FALSE,
  bootstrap_iterations = 100,
  cores = 1,
  model = TRUE,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula". See [stats::lm()] for further details.}

\item{data}{a data frame containing the variables in the model.}

\item{functional}{string containing the functional for which to compute the RIF. Can be one of
"mean", "variance", "quantiles", "gini", or "custom". If "custom"
is selected a \code{custom_rif_function} needs to be provided.}

\item{custom_rif_function}{the RIF function to compute the RIF of the custom functional.
Default is NULL.See examples for further details.}

\item{quantiles}{a vector of length 1 or more with quantile positions to calculate the RIF.
Each quantile is indicated with value between 0 and 1. Only required if \code{functional = "quantiles"}.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{dep_var}.
The default (\code{NULL)} is equivalent to \code{weights = rep(1/nx, nx)},
where nx is the length of (the finite entries of) \code{dep_var}.}

\item{bootstrap}{boolean (Default = FALSE) indicating if bootstrapped standard errors shall be computed}

\item{bootstrap_iterations}{positive integer indicating the number of bootstrap iterations to execute.
Only required if \code{bootstrap = TRUE}.}

\item{cores}{positive integer indicating the number of cores to use when computing bootstrap standard errors.
Only required if \code{bootstrap = TRUE}.}

\item{model}{WHY DO WE NEED THIS PARAMETER?}

\item{...}{additional parameters passed to the \code{custom_rif_function}.
Apart from \code{dep_var} they must have a different name than the the ones of
\code{est_rif}. For instance, if you want to pass weights to the
\code{custom_rif_function}, name them \code{custom_weights}.}
}
\value{
an object of class \code{rifreg} containing the RIF regression estimate,
        bootstrap standard errors, the RIF values and further information.
}
\description{
Estimate the recentered influence function regression (RIFREG) for a chosen
functional of interest.
}
\examples{

example_data <- CPSmen8305[1:300,]
example_weights <- CPSmen8305$weights[1:300]

rifreg <- est_rifreg(formula = log(wage) ~ union + age,
                     data = example_data,
                     functional = "quantiles",
                     custom_rif_function = NULL,
                     quantiles = seq(0.1, 0.9, 0.1),
                     weights = example_weights,
                     bootstrap = TRUE,
                     bootstrap_iterations = 100,
                     cores = 1,
                     model = TRUE)

# custom function
custom_variance_function <- function(dep_var, custom_weights){
  weights <- check_weights(dep_var, weights = custom_weights)
  weighted_mean <- weighted.mean(x = dep_var, w = weights)
  rif <- (dep_var - weighted_mean)^2
  rif <- data.frame(rif, weights)
  names(rif) <- c("rif_variance", "weights")
  return(rif)
}


rifreg <- est_rifreg(
  formula = log(wage) ~ union + age,
  data = example_data,
  functional = "custom",
  custom_rif_function = custom_variance_function,
  quantiles = NULL,
  weights = NULL,
  bootstrap = FALSE,
  cores = 1,
  model = TRUE,
  custom_weights = example_weights)

}
\references{
Firpo, Sergio P., Nicole M. Fortin, and Thomas Lemieux. 2009. “Unconditional Quantile
Regressions.” \emph{Econometrica} 77(3): 953–73.

Cowell, Frank A., and Emmanuel Flachaire. 2015. “Statistical Methods for Distributional Analysis.”
In Anthony B. Atkinson and François Bourguignon (eds.), \emph{Handbook of Income Distribution}. Amsterdam: Elsevier.
}
