% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rif.R
\name{integrate_generalized_lorenz_curve}
\alias{integrate_generalized_lorenz_curve}
\title{Estimate Gini coefficient (slightly adjusted code from Rothe (2015))
Integrate generalized Lorenz curve}
\usage{
integrate_generalized_lorenz_curve(dep_var, weights)
}
\description{
Estimate Gini coefficient (slightly adjusted code from Rothe (2015))
Integrate generalized Lorenz curve
}
