% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rifreg}
\alias{plot.rifreg}
\title{Plot Coefficients for an \code{rifreg} Object}
\usage{
\method{plot}{rifreg}(x, varselect = NULL, confidence_level = 0.05, ...)
}
\arguments{
\item{x}{an object of class "rifreg", usually, a result of a call to [rifreg()] with code{functional = "quantiles"}.}

\item{varselect}{vector of length 1 or more containig the names of the covariates to display.}

\item{confidence_level}{numeric value between 0 and 1 (default = 0.95) that defines the confidence interval
plotted as a ribbon and defined as \code{qnorm(confidence_level/2)} * standard error.
Only required if bootstrapped standard errors where computed.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
a ggplot containing the coefficients for each (selected) covariate
}
\description{
Coefficients are plotted for each quantile and each covariate.
Specific covariates can be selected and SE can also be displayed.
}
\examples{

data <- CPSmen8305[1:300,]
rifreg <- est_rifreg(formula = log(wage) ~ union + age,
                     data = data,
                     functional = "quantiles",
                     custom_functional = NULL,
                     probs = seq(0.1, 0.9, 0.1),
                     weights = weights,
                     bootstrap = FALSE,
                     bootstrap_iterations = 100,
                     cores = 1,
                     model = TRUE)

plot(rifreg)

plot(rifreg, varselect = c("age", "unionyes"), confidence_level = 0.1)

}
