% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rifreg}
\alias{print.rifreg}
\title{print method for class "rifreg"}
\usage{
\method{print}{rifreg}(x, ...)
}
\arguments{
\item{x}{an object of class "rifreg", usually , a result of a call to [rifreg()].}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
the function \code{print.rifreg()} returns the the covariates' coefficients
        of the RIF regressions derived from  the fitted linear model given in object \code{x}.
}
\description{
print method for class "rifreg"
}
\examples{

rifreg <- est_rifreg(formula = log(wage) ~ union +
                                           nonwhite +
                                           married +
                                           education +
                                           experience,
                     data = men8385,
                     statistic = "quantiles",
                     probs = seq(0.1, 0.9, 0.1),
                     weights = weights)

print(rifreg)

}
