% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rifreg}
\alias{summary.rifreg}
\title{summary method for class "rifreg"}
\usage{
\method{summary}{rifreg}(object, ...)
}
\arguments{
\item{object}{an object of class "rifreg", usually, a result of a call to [rifreg()].}

\item{...}{other parameters to be passed through to summary functions.}
}
\value{
the function \code{summary.rifreg()} returns a list of summary statistics derived from
        the rifreg object given in \code{object}. For further details see [stats::summary.lm()].
}
\description{
summary method for class "rifreg"
}
\examples{

data <- CPSmen8305[1:300,]
rifreg <- est_rifreg(formula = log(wage) ~ union + age,
                     data = data,
                     functional = "quantiles",
                     custom_functional = NULL,
                     probs = 0.5,
                     weights = weights,
                     bootstrap = FALSE,
                     bootstrap_iterations = 100,
                     cores = 1,
                     model = TRUE)

summary(rifreg)

}
